Add search functionality to dashboard header (#10)
### TL;DR

Implemented a search page in the dashboard header.

### What changed?

New files were created to setup a search page within the dashboard header. Also, the routes were updated to include the new 'DashboardSearch' route.

### How to test?

Run the dashboard and go to the dashboard header, the search page should be present and accessible.

### Why make this change?

This addition of a search page makes searching within the dashboard easy and accessible, improving the user experience.

---