Update Table View in Dashboard Tasks Page (#17)
### TL;DR
Update the tasks page with a new TableView component and bring in functionality related to Table handling.

### What changed?
Added a new TableView component to the tasks page along with related functions for handling data in a Table format.

### How to test?
Test the functionality of the new TableView component by interacting with the Table view on the tasks page.

### Why make this change?
Introduce a new way to visualize and interact with data in a tabular format for improved user experience and data management.

---